+++ podcasts/NowPlayingPlayerItemViewController.swift
@@ -103,6 +103,10 @@ class NowPlayingPlayerItemViewController: PlayerItemViewController {
             tapGesture.numberOfTapsRequired = 1
             tapGesture.numberOfTouchesRequired = 1
             floatingVideoView.addGestureRecognizer(tapGesture)
+
+            floatingVideoView.onFullScreenTapped = { [weak self] in
+                self?.videoTapped()
+            }
         }
     }
 
@@ -605,13 +609,19 @@ class NowPlayingPlayerItemViewController: PlayerItemViewController {
     }
 
     @objc private func videoTapped() {
-        guard PlaybackManager.shared.currentEpisode != nil else { return }
+        guard PlaybackManager.shared.currentEpisode != nil, presentedViewController == nil else { return }
 
         if PlaybackManager.shared.shouldRenderVideo() {
             let videoController = VideoViewController()
             videoViewController = videoController
-            videoViewController?.modalTransitionStyle = .crossDissolve
             videoViewController?.modalPresentationStyle = .fullScreen
+            if #available(iOS 18.0, *) {
+                videoViewController?.preferredTransition = .zoom { [weak self] _ in
+                    self?.floatingVideoView
+                }
+            } else {
+                videoViewController?.modalTransitionStyle = .crossDissolve
+            }
             videoViewController?.willAttachPlayer = { [weak self] in
                 self?.floatingVideoView.player = nil
             }